+++ projeto_android/Vestlist/app/src/main/java/com/alex/vestlist/source/StudentSourceContract.java
@@ -60,7 +60,7 @@ interface StudentSourceContract {
     public long insertList(ExerciseList list);
     public long insertDoubt(Doubt doubt);
 
-    public long updateDoubt(Doubt doubt);
+    public int updateDoubt(Doubt doubt);
     public int updateList(ExerciseList list);
 
     /**
